rename verification.py -> verifiy.py
